Restructure the carousel library and polish ts types
