+++ services/web/src/components/EmblaCarousel/SlidesPerViewCarousel/css/base.css
