+++ services/web/src/website/MainPage/AboutSection/components/IdentitySection.tsx
@@ -1,7 +1,8 @@
+import React from 'react';
 import identityProposition from '../StaticContent/identityProposition.json';
 import { Button } from '@/components/ui/button.tsx';
 
-const IdentitySection = () => {
+const IdentitySection: React.FC = () => {
     return (
         <div className="mb-40">
             <h1 className="font-mont font-semibold text-3xl text-primary mb-10">Who are we?</h1>
