+++ services/web/src/website/MainPage/AboutSection/components/PastEventSection.tsx
@@ -1,12 +1,13 @@
 import EventTicket from './EventTicket.tsx';
 import pastEvents from '../StaticContent/pastEvents.json';
-import '@/components/EmblaCarousel/css/base.css';
-import '@/components/EmblaCarousel/css/embla.css';
-import EmblaCarousel from '@/components/EmblaCarousel/js/EmblaCarousel.tsx';
+import '@/components/EmblaCarousel/SlidesPerViewCarousel/css/base.css';
+import '@/components/EmblaCarousel/SlidesPerViewCarousel/css/embla.css';
+import EmblaCarousel from '@/components/EmblaCarousel/SlidesPerViewCarousel/js/EmblaCarousel.tsx';
 import { EmblaOptionsType } from 'embla-carousel';
+import React from 'react';
 
-const PastEventSection = () => {
-    const SLIDES = pastEvents.map((pastEvent, index) => (
+const PastEventSection: React.FC = () => {
+    const SLIDES: React.ReactElement[] = pastEvents.map((pastEvent, index) => (
         <EventTicket imgSrc={pastEvent.cover_picture} title={pastEvent.title} tags={pastEvent.tags} key={index} />
     ));
